refactor(services): Reorganize services
