+++ src/app/app.component.spec.ts
@@ -1,15 +1,15 @@
 import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { LanguageService } from './core/language/language.service';
 
 import { TestModule } from './test.module';
+
+import { LanguageService } from './core/language/language.service';
+
+import { NavigatorModule, NavigatorRoutingModule } from './pages';
+
 import { AppComponent } from './app.component';
 import { SpinnerComponent } from './core/spinner/spinner.component';
-import { NavigatorModule, NavigatorRoutingModule } from './pages';
-import { MediaService } from './core/media.service';
-import { RequestService } from './core/request.service';
-import { LoggingService } from './core/logging.service';
-import { provideAppStore } from './core/core.module';
+
 
 import {} from 'jasmine';
 
@@ -27,11 +27,7 @@ describe('AppComponent', () => {
         SpinnerComponent
       ],
       providers: [
-        LanguageService,
-        provideAppStore(),
-        MediaService,
-        LoggingService,
-        RequestService
+        LanguageService
       ]
     });
     TestBed.compileComponents();
