+++ src/app/context/context-item/context-item.component.spec.ts
@@ -2,6 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TestModule } from '../../test.module';
 import { SharedModule } from '../../shared/shared.module';
+
 import { ContextItemComponent } from './context-item.component';
 
 describe('ContextItemComponent', () => {
