+++ src/app/context/context-tool/context-tool.component.spec.ts
@@ -2,13 +2,13 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TestModule } from '../../test.module';
 import { SharedModule } from '../../shared/shared.module';
+
+import { ToolService } from '../../tool/shared/tool.service';
+
 import { ContextToolComponent } from './context-tool.component';
 import { ContextItemComponent } from '../context-item/context-item.component';
-import { ContextService } from '../../core/context.service';
-import { RequestService } from '../../core/request.service';
-import { LoggingService } from '../../core/logging.service';
-import { ToolService } from '../../core/tool.service';
-import { provideAppStore } from '../../core/core.module';
+import { ContextService } from '../shared/context.service';
+
 
 describe('ContextToolComponent', () => {
   let component: ContextToolComponent;
@@ -26,10 +26,7 @@ describe('ContextToolComponent', () => {
       ],
       providers: [
         ContextService,
-        RequestService,
-        LoggingService,
-        ToolService,
-        provideAppStore()
+        ToolService
       ]
     })
     .compileComponents();
