+++ src/app/context/context.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 
+import { SharedModule } from '../shared/shared.module';
+
 import { ContextToolComponent } from './context-tool/context-tool.component';
 import { ContextItemComponent } from './context-item/context-item.component';
+import { ContextService } from './shared/context.service';
 
-import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   imports: [
@@ -13,6 +15,9 @@ import { SharedModule } from '../shared/shared.module';
     ContextToolComponent,
     ContextItemComponent
   ],
-  entryComponents: [ContextToolComponent]
+  entryComponents: [ContextToolComponent],
+  providers: [
+    ContextService
+  ]
 })
 export class ContextModule { }
