+++ src/app/core/logging.service.spec.ts
@@ -1,6 +1,4 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
+import { TestBed, inject } from '@angular/core/testing';
 import { LoggingService } from './logging.service';
 
 describe('LoggingService', () => {
