+++ src/app/pages/navigator/navigator.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../../shared/shared.module';
-import { MapModule } from '../../map/map.module';
 import { ContextModule } from '../../context/context.module';
+import { MapModule } from '../../map/map.module';
 import { SearchModule } from '../../search/search.module';
 import { ToolModule } from '../../tool/tool.module';
