+++ src/app/search/shared/search-source.service.spec.ts
@@ -1,11 +1,11 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { HttpModule, JsonpModule } from '@angular/http';
-import { SearchSourceService } from './search-source.service';
 
+import { SearchSourceService } from './search-source.service';
 import {
-   provideSearchSource,
-   provideSearchSourceService
-} from './core.module';
+  provideSearchSources,
+  provideSearchSourceService
+} from '../search.module';
 
 describe('SearchSourceService', () => {
   beforeEach(() => {
@@ -16,7 +16,7 @@ describe('SearchSourceService', () => {
       ],
       providers: [
         provideSearchSourceService(),
-        provideSearchSource()
+        ...provideSearchSources()
       ]
     });
   });
