+++ src/app/store/reducers/contexts-list.reducer.ts
@@ -1,4 +1,4 @@
-import { Context } from '../context/shared/context.interface';
+import { Context } from '../../context/shared/context.interface';
 
 export function contextsList(state: Context[] = [], {type, payload}) {
   switch (type) {
