+++ src/app/store/reducers/map-layers.reducer.ts
@@ -1,4 +1,4 @@
-import { LayerOptions } from '../map/shared/layers/layer';
+import { LayerOptions } from '../../map/shared/layers/layer';
 
 export function mapLayers(state: Array<LayerOptions> = [], {type, payload}) {
   switch (type) {
