+++ src/app/store/reducers/map-view.reducer.ts
