+++ src/app/store/reducers/search-results.reducer.ts
@@ -1,4 +1,4 @@
-import { SearchResult } from '../search/shared/search-result.interface';
+import { SearchResult } from '../../search/shared/search-result.interface';
 
 export function searchResults(state: SearchResult[] = [], {type, payload}) {
   switch (type) {
