+++ src/app/store/reducers/tools-list.reducer.ts
@@ -1,4 +1,4 @@
-import { Tool } from '../tool/shared/tool.interface';
+import { Tool } from '../../tool/shared/tool.interface';
 
 export function toolsList(state: Tool[] = [], {type, payload}) {
   switch (type) {
