+++ src/app/test.module.ts
@@ -1,7 +1,14 @@
 import { NgModule } from '@angular/core';
 import { MaterialModule } from '@angular/material';
-import { HttpModule, Http } from '@angular/http';
+import { HttpModule, Http, JsonpModule } from '@angular/http';
 import { TranslateModule, TranslateLoader, TranslateStaticLoader } from 'ng2-translate';
+
+import { MediaService } from './core/media.service';
+import { LoggingService } from './core/logging.service';
+import { RequestService } from './core/request.service';
+
+import { provideIgoStore } from './store/store.module';
+
 
 export function createTranslateLoader(http: Http) {
   return new TranslateStaticLoader(http, './assets/locale', '.json');
@@ -10,6 +17,7 @@ export function createTranslateLoader(http: Http) {
 @NgModule({
   imports: [
     HttpModule,
+    JsonpModule,
     TranslateModule.forRoot({
       provide: TranslateLoader,
       useFactory: createTranslateLoader,
@@ -21,6 +29,12 @@ export function createTranslateLoader(http: Http) {
     HttpModule,
     MaterialModule,
     TranslateModule
+  ],
+  providers: [
+    MediaService,
+    LoggingService,
+    RequestService,
+    provideIgoStore()
   ]
 })
