+++ src/app/tool/shared/tool.service.spec.ts
@@ -1,4 +1,5 @@
 import { TestBed, inject } from '@angular/core/testing';
+
 import { ToolService } from './tool.service';
 
 describe('ToolService', () => {
