+++ src/app/tool/tool.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../shared/shared.module';
+
+import { ToolService } from './shared/tool.service';
 import { ToolbarComponent } from './toolbar/toolbar.component';
 import { ToolbarItemComponent } from './toolbar-item/toolbar-item.component';
 import { ToolboxComponent } from './toolbox/toolbox.component';
@@ -17,6 +19,9 @@ import { ToolboxComponent } from './toolbox/toolbox.component';
     ToolbarComponent,
     ToolbarItemComponent,
     ToolboxComponent
+  ],
+  providers: [
+    ToolService
   ]
 })
 export class ToolModule { }
